st/texture-cache: Switch to integers for resource scales
Since clutter_actor_get_resource_scale() now returns an unsigned integer
instead of a float for the scale, switch to ints for storing that scale
in the texture cache, too.

https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/1287

+++ src/st/st-texture-cache.h
@@ -60,7 +60,7 @@ st_texture_cache_load_sliced_image (StTextureCache *cache,
                                     gint            grid_width,
                                     gint            grid_height,
                                     gint            paint_scale,
-                                    gfloat          resource_scale,
+                                    unsigned int    resource_scale,
                                     GFunc           load_callback,
                                     gpointer        user_data);
 
@@ -74,24 +74,24 @@ ClutterActor *st_texture_cache_load_gicon (StTextureCache *cache,
                                            GIcon          *icon,
                                            gint            size,
                                            gint            paint_scale,
-                                           gfloat          resource_scale);
+                                           unsigned int    resource_scale);
 
 ClutterActor *st_texture_cache_load_file_async (StTextureCache    *cache,
                                                 GFile             *file,
                                                 int                available_width,
                                                 int                available_height,
                                                 int                paint_scale,
-                                                gfloat             resource_scale);
+                                                unsigned int       resource_scale);
 
 CoglTexture     *st_texture_cache_load_file_to_cogl_texture (StTextureCache *cache,
                                                              GFile          *file,
                                                              gint            paint_scale,
-                                                             gfloat          resource_scale);
+                                                             unsigned int    resource_scale);
 
 cairo_surface_t *st_texture_cache_load_file_to_cairo_surface (StTextureCache *cache,
                                                               GFile          *file,
                                                               gint            paint_scale,
-                                                              gfloat          resource_scale);
+                                                              unsigned int    resource_scale);
 
 /**
  * StTextureCacheLoader: (skip)
